requirements.txt
django
django_rest_framework
django-cors-headers
django-extras

